feat: unify facilitator picker - extract shared FacilitatorSelect component
- New facilitator-select.tsx: searchable combobox with current-on-top, no auto-focus
- facilitator-picker.tsx: thin wrapper adding server action + toast
- session-form.tsx: uses same FacilitatorSelect with fullWidth mode
- Both show UserRoundPlus icon consistently
